Move ImageService to Core module
