+++ src/app/core/image.service.ts
@@ -8,7 +8,6 @@ const routes = {
 };
 
 export interface RandomQuoteContext {
-  // The quote's category: 'dev', 'explicit'...
   category: string;
 }
