+++ src/app/shell/header/header.component.spec.ts
@@ -1,5 +1,4 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { ThemeModule } from '../../@theme/theme.module';
@@ -12,12 +11,7 @@ describe('HeaderComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: 
-      [
-        ThemeModule,
-        RouterTestingModule, 
-        NgbModule 
-      ],
+      imports: [ ThemeModule, RouterTestingModule ],
       declarations: [HeaderComponent],
       providers: [{ provide: AuthenticationService, useClass: MockAuthenticationService }]
     }).compileComponents();
